refactor: split MCP tools into package modules
Move search_portal, get_document, and run_code into dedicated modules under src/okp_mcp/tools while preserving the okp_mcp.tools import surface. Update tests and docs so FastMCP tool registration and helper patch targets continue to work after the package split.


stack-info: PR: https://github.com/rhel-lightspeed/okp-mcp/pull/120, branch: major/stack/1
