+++ src/okp_mcp/tools/__init__.py
@@ -1,0 +1,25 @@
+"""MCP tool definitions for RHEL OKP knowledge base search."""
+
+from .document import (
+    _doc_id_filter,
+    _escape_solr_phrase,
+    _fetch_document_raw,
+    _fetch_document_with_query,
+    _format_document,
+    _normalize_doc_id,
+    get_document,
+)
+from .run_code import run_code
+from .search import search_portal
+
+__all__ = [
+    "_doc_id_filter",
+    "_escape_solr_phrase",
+    "_fetch_document_raw",
+    "_fetch_document_with_query",
+    "_format_document",
+    "_normalize_doc_id",
+    "get_document",
+    "run_code",
+    "search_portal",
+]
