+++ src/okp_mcp/tools/shared.py
@@ -1,0 +1,7 @@
+"""Shared constants and helpers for MCP tool modules."""
+
+DOCUMENT_FL = (
+    "id,allTitle,title,heading_h1,main_content,view_uri,documentKind,"
+    "product,documentation_version,lastModifiedDate,"
+    "cve_details,portal_synopsis,portal_summary"
+)
